+++ src/Lawn/ZenGarden.cpp
@@ -1,3 +1,4 @@
+#include <cstdint>
 #include <time.h>
 #include "Plant.h"
 #include "Board.h"
@@ -499,7 +500,7 @@ void ZenGarden::MouseDownWithMoneySign(Plant* thePlant)
 void ZenGarden::PlantFertilized(Plant* thePlant)
 {
     PottedPlant* aPottedPlant = PottedPlantFromIndex(thePlant->mPottedPlantIndex);
-    aPottedPlant->mLastFertilizedTime = (int64_t)time(0);
+    aPottedPlant->mLastFertilizedTime = static_cast<int64_t>(time(0));
     aPottedPlant->mPlantAge = (PottedPlantAge)((int)aPottedPlant->mPlantAge + 1);
     aPottedPlant->mPlantNeed = PottedPlantNeed::PLANTNEED_NONE;
     aPottedPlant->mTimesFed = 0;
@@ -544,7 +545,7 @@ void ZenGarden::PlantFertilized(Plant* thePlant)
 void ZenGarden::PlantFulfillNeed(Plant* thePlant)
 {
     PottedPlant* aPottedPlant = PottedPlantFromIndex(thePlant->mPottedPlantIndex);
-    aPottedPlant->mLastNeedFulfilledTime = (int64_t)time(0);
+    aPottedPlant->mLastNeedFulfilledTime = static_cast<int64_t>(time(0));
     aPottedPlant->mPlantNeed = PottedPlantNeed::PLANTNEED_NONE;
     aPottedPlant->mTimesFed = 0;
 
@@ -623,7 +624,7 @@ void ZenGarden::PlantWatered(Plant* thePlant)
     {
         aTimeSpan = 9;
     }
-    aPottedPlant->mLastWateredTime = (int64_t)time(0) - aTimeSpan;
+    aPottedPlant->mLastWateredTime = static_cast<int64_t>(time(0)) - aTimeSpan;
 
     mApp->PlayFoley(FoleyType::FOLEY_SPAWN_SUN);
     mBoard->AddCoin(thePlant->mX + 40, thePlant->mY, CoinType::COIN_SILVER, CoinMotion::COIN_MOTION_COIN);
@@ -745,7 +746,7 @@ void ZenGarden::RemoveHappyEffect(Plant* thePlant)
 bool ZenGarden::WasPlantNeedFulfilledToday(PottedPlant* thePottedPlant)
 {
     time_t aNowTime = time(0);
-    int64_t aNow = (int64_t)aNowTime;
+    int64_t aNow = static_cast<int64_t>(aNowTime);
     if (aNow - thePottedPlant->mLastNeedFulfilledTime < 3600)
     {
         return true;
@@ -761,7 +762,7 @@ bool ZenGarden::WasPlantNeedFulfilledToday(PottedPlant* thePottedPlant)
 bool ZenGarden::PlantShouldRefreshNeed(PottedPlant* thePottedPlant)
 {
     time_t aNowTime = time(0);
-    int64_t aNow = (int64_t)aNowTime;
+    int64_t aNow = static_cast<int64_t>(aNowTime);
     if (aNow - thePottedPlant->mLastWateredTime < 3600)
     {
         return false;
@@ -783,7 +784,7 @@ void ZenGarden::RefreshPlantNeeds(PottedPlant* thePottedPlant)
 
     if (Plant::IsAquatic(thePottedPlant->mSeedType))
     {
-        thePottedPlant->mLastWateredTime = (int64_t)time(0);
+        thePottedPlant->mLastWateredTime = static_cast<int64_t>(time(0));
         thePottedPlant->mPlantNeed = (PottedPlantNeed)RandRangeInt((int)PottedPlantNeed::PLANTNEED_BUGSPRAY, (int)PottedPlantNeed::PLANTNEED_PHONOGRAPH);
     }
     else
@@ -810,7 +811,7 @@ void ZenGarden::UpdatePlantNeeds()
 
 bool ZenGarden::WasPlantFertilizedInLastHour(PottedPlant* thePottedPlant)
 {
-    return (int64_t)time(0) - thePottedPlant->mLastFertilizedTime < 3600;
+    return static_cast<int64_t>(time(0)) - thePottedPlant->mLastFertilizedTime < 3600;
 }
 
 //0x51EA30
@@ -827,7 +828,7 @@ PottedPlantNeed ZenGarden::GetPlantsNeed(PottedPlant* thePottedPlant)
         return PottedPlantNeed::PLANTNEED_NONE;
     }
 
-    int64_t aNow = (int64_t)time(0);
+    int64_t aNow = static_cast<int64_t>(time(0));
     bool aTooLongSinceWatering = aNow - thePottedPlant->mLastWateredTime > 15;
     bool aTooShortSinceWatering = aNow - thePottedPlant->mLastWateredTime < 3;
 
@@ -906,7 +907,7 @@ void ZenGarden::MouseDownWithFeedingTool(int x, int y, CursorType theCursorType)
         {
             WakeStinky();
             mApp->AddTodParticle(aStinky->mPosX + 40.0f, aStinky->mPosY + 40.0f, aStinky->mRenderOrder + 1, ParticleEffect::PARTICLE_PRESENT_PICKUP);
-            mApp->mPlayerInfo->mLastStinkyChocolateTime = time(0);
+            mApp->mPlayerInfo->mLastStinkyChocolateTime = static_cast<int32_t>(time(0));
             mApp->mPlayerInfo->mPurchases[(int)StoreItem::STORE_ITEM_CHOCOLATE]--;
 
             mApp->PlayFoley(FoleyType::FOLEY_WAKEUP);
@@ -992,7 +993,7 @@ void ZenGarden::MouseDownWithFeedingTool(int x, int y, CursorType theCursorType)
 void ZenGarden::FeedChocolateToPlant(Plant* thePlant)
 {
     PottedPlant* aPottedPlant = PottedPlantFromIndex(thePlant->mPottedPlantIndex);
-    aPottedPlant->mLastChocolateTime = (int64_t)time(0);
+    aPottedPlant->mLastChocolateTime = static_cast<int64_t>(time(0));
     thePlant->mLaunchCounter = 60;
     mApp->AddTodParticle(thePlant->mX + 40.0f, thePlant->mY + 40.0f, thePlant->mRenderOrder + 1, ParticleEffect::PARTICLE_PRESENT_PICKUP);
 }
@@ -1279,7 +1280,7 @@ void ZenGarden::AddStinky()
     if (!mApp->mPlayerInfo->mHasSeenStinky)
     {
         mApp->mPlayerInfo->mHasSeenStinky = 1;
-        mApp->mPlayerInfo->mPurchases[(int)StoreItem::STORE_ITEM_STINKY_THE_SNAIL] = time(0);
+        mApp->mPlayerInfo->mPurchases[(int)StoreItem::STORE_ITEM_STINKY_THE_SNAIL] = static_cast<int32_t>(time(0));
     }
 
     GridItem* aStinky = mBoard->mGridItems.DataArrayAlloc();
@@ -2243,7 +2244,7 @@ void ZenGarden::SetPlantAnimSpeed(Plant* thePlant)
 int ZenGarden::PlantGetMinutesSinceHappy(Plant* thePlant)
 {
     PottedPlant* aPottedPlant = PottedPlantFromIndex(thePlant->mPottedPlantIndex);
-    int aMinutes = (int)(((int64_t)time(0) - aPottedPlant->mLastNeedFulfilledTime) / 60);
+    int aMinutes = static_cast<int>((static_cast<int64_t>(time(0)) - aPottedPlant->mLastNeedFulfilledTime) / 60);
     if (PlantHighOnChocolate(aPottedPlant))
     {
         aMinutes = 0;
@@ -2290,7 +2291,7 @@ void ZenGarden::ResetPlantTimers(PottedPlant* thePottedPlant)
 void ZenGarden::PottedPlantUpdate(Plant* thePlant)
 {
     PottedPlant* aPottedPlant = PottedPlantFromIndex(thePlant->mPottedPlantIndex);
-    int64_t aNow = (int64_t)time(0);
+    int64_t aNow = static_cast<int64_t>(time(0));
     if (aPottedPlant->mLastWateredTime > aNow || 
         aPottedPlant->mLastNeedFulfilledTime > aNow || 
         aPottedPlant->mLastFertilizedTime > aNow || 
@@ -2356,7 +2357,7 @@ void ZenGarden::DrawPlantOverlay(Graphics* g, Plant* thePlant)
 //0x521FE0
 void ZenGarden::WakeStinky()
 {
-    mApp->mPlayerInfo->mPurchases[(int)StoreItem::STORE_ITEM_STINKY_THE_SNAIL] = time(0);
+    mApp->mPlayerInfo->mPurchases[(int)StoreItem::STORE_ITEM_STINKY_THE_SNAIL] = static_cast<int32_t>(time(0));
     mApp->PlaySample(SOUND_TAP);
     mBoard->ClearAdvice(AdviceType::ADVICE_STINKY_SLEEPING);
     gLawnApp->mPlayerInfo->mHasWokenStinky = 1;
@@ -2370,7 +2371,7 @@ bool ZenGarden::IsStinkyHighOnChocolate()
 
 bool ZenGarden::PlantHighOnChocolate(PottedPlant* thePottedPlant)
 {
-    return (int64_t)time(0) - thePottedPlant->mLastChocolateTime < 300;
+    return static_cast<int64_t>(time(0)) - thePottedPlant->mLastChocolateTime < 300;
 }
 
 bool ZenGarden::IsStinkySleeping()
